// src/types/raw/RawUHC.ts
export interface RawUHC {
    clearup_achievement?: boolean;
    coins?: number;
    deaths?: number;
    deaths_brawl?: number;
    'deaths_duo brawl'?: number;
    deaths_duo_brawl?: number;
    'deaths_no diamonds'?: number;
    'deaths_red vs blue'?: number;
    deaths_solo?: number;
    'deaths_solo brawl'?: number;
    deaths_solo_brawl?: number;
    equippedKit?: string;
    extra_ultimates_crafted?: number;
    extra_ultimates_crafted_duo_brawl?: number;
    extra_ultimates_crafted_solo?: number;
    extra_ultimates_crafted_solo_brawl?: number;
    heads_eaten?: number;
    heads_eaten_brawl?: number;
    'heads_eaten_duo brawl'?: number;
    heads_eaten_duo_brawl?: number;
    'heads_eaten_no diamonds'?: number;
    'heads_eaten_red vs blue'?: number;
    heads_eaten_solo?: number;
    'heads_eaten_solo brawl'?: number;
    heads_eaten_solo_brawl?: number;
    kills?: number;
    kills_brawl?: number;
    'kills_duo brawl'?: number;
    kills_duo_brawl?: number;
    'kills_no diamonds'?: number;
    'kills_red vs blue'?: number;
    kills_solo?: number;
    'kills_solo brawl'?: number;
    kills_solo_brawl?: number;
    kills_solo2?: number;
    kills2?: number;
    kit_ARCHERY_TOOLS?: number;
    kit_ECOLOGIST?: number;
    kit_FARMER?: number;
    kit_HORSEMAN?: number;
    kit_LEATHER_ARMOR?: number;
    kit_LOOTER?: number;
    kit_LUNCH_BOX?: number;
    kit_MAGIC_TOOLS?: number;
    kit_TRAPPER?: number;
    kit_WORKING_TOOLS?: number;
    packages?: string[];
    perk_alchemy_line_a?: number;
    perk_alchemy_line_b?: number;
    perk_alchemy_line_c?: number;
    perk_alchemy_prestige?: number;
    perk_apprentice_line_a?: number;
    perk_apprentice_line_b?: number;
    perk_apprentice_line_c?: number;
    perk_apprentice_prestige?: number
    perk_armorsmith_line_a?: number;
    perk_armorsmith_line_b?: number;
    perk_armorsmith_line_c?: number;
    perk_armorsmith_prestige?: number;
    perk_bloodcraft_line_a?: number;
    perk_bloodcraft_line_b?: number;
    perk_bloodcraft_line_c?: number;
    perk_bloodcraft_prestige?: number;
    perk_cooking_line_a?: number;
    perk_cooking_line_b?: number;
    perk_cooking_line_c?: number;
    perk_cooking_prestige?: number;
    perk_enchanting_line_a?: number;
    perk_enchanting_line_b?: number;
    perk_enchanting_line_c?: number;
    perk_enchanting_prestige?: number;
    perk_engineering_line_a?: number;
    perk_engineering_line_b?: number;
    perk_engineering_line_c?: number;
    perk_engineering_prestige?: number;
    perk_hunter_line_a?: number;
    perk_hunter_line_b?: number;
    perk_hunter_line_c?: number;
    perk_hunter_prestige?: number;
    perk_invention_line_a?: number;
    perk_invention_line_b?: number;
    perk_invention_line_c?: number;
    perk_invention_prestige?: number;
    perk_strategist_line_a?: number;
    perk_strategist_line_b?: number;
    perk_strategist_line_c?: number;
    perk_strategist_prestige?: number;
    perk_survivalism_line_a?: number;
    perk_survivalism_line_b?: number;
    perk_survivalism_line_c?: number;
    perk_survivalism_prestige?: number;
    perk_toolsmithing_line_a?: number;
    perk_toolsmithing_line_b?: number;
    perk_toolsmithing_line_c?: number;
    perk_toolsmithing_prestige?: number;
    perk_weaponsmith_line_a?: number;
    perk_weaponsmith_line_b?: number;
    perk_weaponsmith_line_c?: number;
    perk_weaponsmith_prestige?: number;
    saved_stats?: boolean;
    score?: number;
    teammate_damage?: boolean;
    uhc_parkour_1?: boolean;
    uhc_parkour_2?: boolean;
    ultimates_crafted?: number;
    ultimates_crafted_duo_brawl?: number;
    ultimates_crafted_solo?: number;
    ultimates_crafted_solo_brawl?: number;
    wins?: number;
    wins_brawl?: number;
    'wins_duo brawl'?: number;
    wins_duo_brawl?: number;
    'wins_no diamonds'?: number;
    'wins_red vs blue'?: number;
    wins_solo?: number;
    wins_solo_brawl?: number;
    wins_solo2?: number;
    wins2?: number;
}